requirements.txt
numpy == 1.21.2
google-cloud-storage == 1.42.0